% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{list_job}
\alias{list_job}
\title{Return a list of your jobs.}
\usage{
list_job(client)
}
\arguments{
\item{client}{Treasure Data client generated by td function.

data.frame containing the number of jobs, to, from, and jobs an array of the jobs.}
}
\description{
Return a list of your jobs.
}

