% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{list_table}
\alias{list_table}
\title{Returns a list of your tables.}
\usage{
list_table(client, database = NULL)
}
\arguments{
\item{client}{Treasure Data client generated by td function.}

\item{database(optional):}{name of the database}
}
\description{
Returns a list of your tables.
}

