% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td-job.R
\name{result}
\alias{result}
\title{Returns the result of a specific job.}
\usage{
result(job)
}
\arguments{
\item{job}{tdjob object generated by td_job()}
}
\description{
Before issuing this command, please confirm that the job has been completed successfully via show()
The resulting logs can be large, so using a large timeout value (i.e. several minutes) is recommended when using this command.
}

