% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td-job.R
\docType{methods}
\name{show,tdjob-method}
\alias{show,tdjob-method}
\title{Shows the status and logs of a specific job.}
\usage{
\S4method{show}{tdjob}(object)
}
\arguments{
\item{job}{tdjob object generated by td_job()}
}
\description{
Shows the status and logs of a specific job.
}

