% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td-job.R
\name{status}
\alias{status}
\title{Shows the status of a specific job.}
\usage{
status(job)
}
\arguments{
\item{job}{tdjob object generated by td_job()}
}
\description{
It is faster and more robust than show()
}

