% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{table_swap}
\alias{table_swap}
\title{Swaps the contents of two tables.}
\usage{
table_swap(client, table1, table2, database = NULL)
}
\arguments{
\item{client}{Treasure Data client generated by td function.}

\item{table1:}{table name (before)}

\item{table2:}{table name (after)}

\item{database(optional):}{name of the database}
}
\description{
Swaps the contents of two tables.
}

