% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdclientr.R
\docType{package}
\name{tdclientr}
\alias{tdclientr}
\alias{tdclientr-package}
\alias{version}
\title{tdclientr}
\format{An object of class \code{character} of length 1.}
\usage{
version
}
\description{
Provides an Access to Treasure Data via R
}
\keyword{datasets}

